feat(maven-plugin): rename 'lint' goal to 'check' 'lint' is ambiguous since it may imply autofix. 'check' is was is used e.g. by checkstyle, and is explicit The 'lint' goal is still supported, but is deprecated and will be removed in a future release
