+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/CheckMojoIT/detect_and_parse_all_changelogs/pom.xml
@@ -10,6 +10,14 @@
   <artifactId>liquibase-linter-maven-plugin-test</artifactId>
   <version>0.0.1-SNAPSHOT</version>
 
+  <dependencies>
+    <dependency>
+      <groupId>org.springframework.batch</groupId>
+      <artifactId>spring-batch-core</artifactId>
+      <version>5.2.1</version>
+    </dependency>
+  </dependencies>
+
   <build>
     <plugins>
       <plugin>
@@ -18,13 +26,12 @@
         <version>@project.version@</version>
         <configuration>
           <changeLogFile>src/main/resources/config/liquibase/master.xml</changeLogFile>
-          <configurationFile>src/test/resources/liquibase-linter/lqlint.json</configurationFile>
         </configuration>
         <executions>
           <execution>
             <id>default</id>
             <goals>
-              <goal>lint</goal>
+              <goal>check</goal>
             </goals>
           </execution>
         </executions>
