+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/CheckMojoIT/detect_and_parse_all_changelogs/src/main/resources/config/liquibase/changelog/002_create_index_example.xml
