+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/CheckMojoIT/detect_and_parse_all_changelogs/src/main/resources/config/liquibase/changelog/003_create_spring_batch_schema.xml
