+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/CheckMojoIT/detect_lint_issue_with_configuration_using_absolute_path/src/test/resources/liquibase-linter/lqlint.json
