+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/CheckMojoIT/detect_lint_issue_with_configuration_using_relative_path/src/test/resources/liquibase-linter/lqlint.json
@@ -1,0 +1,10 @@
+{
+  "fail-fast": false,
+  "rules": {
+    "index-name": {
+      "pattern": "^{{value}}_IX\\d$\\d$",
+      "dynamicValue": "tableName",
+      "errorMessage": "Index names must be the table name, suffixed with 'IX' and a two-digit number, e.g. FOO_IX01"
+    }
+  }
+}
