+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/CheckMojoIT/detect_lint_issue_with_default_configuration_file/pom.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<project
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xmlns="http://maven.apache.org/POM/4.0.0"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+>
+  <modelVersion>4.0.0</modelVersion>
+
+  <groupId>io.github.liquibase-linter</groupId>
+  <artifactId>liquibase-linter-maven-plugin-test</artifactId>
+  <version>0.0.1-SNAPSHOT</version>
+
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>io.github.liquibase-linter</groupId>
+        <artifactId>liquibase-linter-maven-plugin</artifactId>
+        <version>@project.version@</version>
+        <configuration>
+          <changeLogFile>src/main/resources/config/liquibase/master.xml</changeLogFile>
+        </configuration>
+        <executions>
+          <execution>
+            <id>default</id>
+            <goals>
+              <goal>check</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+    </plugins>
+  </build>
+</project>
