+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/LintMojoIT/support_former_lint_goal/src/main/resources/config/liquibase/master.xml
