+++ liquibase-linter-maven-plugin/src/test/resources-its/io/github/liquibaselinter/mavenplugin/LintMojoIT/support_former_lint_goal/src/test/resources/lqlint.json
